cancer prediction/requirements.txt
Flask
pymongo
joblib
numpy
werkzeug
python-dotenv
